Setting souce for poetry to work Renamed "csv2docx" to "src" now code looks clean
