+++ pyproject.toml
@@ -16,7 +16,10 @@ click = "^7.1.2"
 [tool.poetry.dev-dependencies]
 
 [tool.poetry.scripts]
-convert = "csv2docx.cli:main"
+convert = "cli:main"
+
+[tool.coverage.paths]
+source = ["src", "*/site-packages"]
 
 [build-system]
 requires = ["poetry>=0.12"]
